requirements.txt
appnope==0.1.2
asgiref==3.3.4
backcall==0.2.0
blessings==1.7
boto3==1.17.43
botocore==1.20.64
click==7.1.2
coverage==5.5
decorator==5.0.7
dj-database-url==0.5.0
Django==3.1.7
django-configurations==2.2
django-filter==2.4.0
django-model-utils==4.1.1
django-nose==1.4.7
django-storages==1.11.1
django-unique-upload==0.2.1
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.1.2
flake8==3.9.0
future==0.18.2
gunicorn==20.1.0
ipdb==0.13.7
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
livereload==2.6.3
lunr==0.5.8
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
mkdocs==1.1.2
mkdocs-material==7.1.3
mkdocs-material-extensions==1.0.1
mock==4.0.3
newrelic==6.2.0.156
nltk==3.6.2
nose==1.3.7
nose-progressive==1.5.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
pymdown-extensions==8.1.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
s3transfer==0.3.7
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
